refactor: to build the component
(cherry picked from commit d234bca95aee320e1142beabe50a908a9c99b835)
(cherry picked from commit 1a9c358bdee90f04d42dd55b379897023b4a6c52)
