+++ packages/react-components/src/components/QrCode/QrCode.tsx
