+++ packages/react-components/src/components/QrCode/index.ts
@@ -1,0 +1,1 @@
+export { QRCode } from './QrCode.js';
