+++ packages/react-components/src/components/QrCodeDocumentWrapper/index.ts
@@ -1,0 +1,1 @@
+export { QrCodeDocumentWrapper } from './QrCodeDocumentWrapper.js';
