--- packages/react-components/src/components/qrCode/index.ts
@@ -1,1 +1,0 @@
-export { QRCode } from './qrCode';
